+++ count_number_of_all_words/includes/merging/merge_maps.h
@@ -9,7 +9,7 @@
 #include <map>
 #include "../../includes/queues/tqueue.h"
 
-void merge_maps_queue(t_queue<std::map<std::string, int>> &queue, uint8_t num_of_threads);
+void merge_maps_queue(t_queue<std::map<std::string, size_t>> &queue, uint8_t num_of_threads);
 
 
 #endif //COUNT_NUMBER_OF_ALL_WORDS_MERGE_MAPS_H
