+++ count_number_of_all_words/src/counting/linear_program.cpp
@@ -13,7 +13,7 @@
 
 void linear_count(const std::string &input_filename, const std::string &output_filename_a,
                   const std::string &output_filename_n) {
-    std::map<std::string, int> map_of_words;
+    std::map<std::string, size_t> map_of_words;
     std::vector<std::string> data;
     std::string word;
     read_input_file(input_filename, data);
