+++ count_number_of_all_words/src/files/file_interaface.cpp
@@ -7,7 +7,7 @@
 #include "../../includes/files/file_interface.h"
 #include "../../includes/merging/map_helpers.h"
 
-void print(const std::map<std::string, int> &map_of_words, const std::string &output_filename_a,
+void print(const std::map<std::string, size_t> &map_of_words, const std::string &output_filename_a,
            const std::string &output_filename_n) {
     std::ofstream outfile_alpha;
     std::ofstream outfile_number;
